requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
greenlet==1.1.2
idna==3.3
python-dotenv==0.20.0
requests==2.28.1
SQLAlchemy==1.4.39
urllib3==1.26.11
vk-api==11.9.9
pytest~=7.1.2
vk~=3.0
selenium~=4.3.0